Fix Android

+++ Mlib/Geometry/Mesh/Farthest_Distances.cpp
@@ -1,6 +1,8 @@
 #include "Farthest_Distances.hpp"
 #include <Mlib/Geometry/Mesh/IIntersectable_Mesh.hpp>
 #include <Mlib/Geometry/Plane_Nd.hpp>
+#include <Mlib/Math/Orderable_Fixed_Array.hpp>
+#include <limits>
 
 using namespace Mlib;
 
@@ -9,8 +11,8 @@ VertexDistances Mlib::get_farthest_distances(
     const PlaneNd<double, 3>& plane)
 {
     VertexDistances res{
-        .min = INFINITY,
-        .max = -INFINITY
+        .min = (double)INFINITY,
+        .max = -(double)INFINITY
     };
     for (const auto& v : mesh.get_vertices()) {
         auto dist = dot0d(v, plane.normal) + plane.intercept;
